+++ test/ssl_test.rb
@@ -30,38 +30,6 @@ test "https request" do |app|
   assert_equal "secure", app.res.body
 end
 
-test "hsts header" do |app|
-  Tynn.helpers(Tynn::SSL)
-
-  Tynn.define do
-  end
-
-  app = Tynn::Test.new
-  app.get("/", {}, "HTTPS" => "on")
-
-  header = app.res.headers["Strict-Transport-Security"]
-
-  assert_equal "max-age=15552000; includeSubdomains", header
-end
-
-test "hsts header options" do |app|
-  Tynn.helpers(Tynn::SSL, hsts: {
-    expires: 1,
-    subdomains: false,
-    preload: true
-  })
-
-  Tynn.define do
-  end
-
-  app = Tynn::Test.new
-  app.get("/", {}, "HTTPS" => "on")
-
-  header = app.res.headers["Strict-Transport-Security"]
-
-  assert_equal "max-age=1; preload", header
-end
-
 test "secure cookies" do |app|
   Tynn.helpers(Tynn::SSL)